+++ zanshinsdk/client.py
@@ -117,6 +117,10 @@ class ScanTargetAZURE(dict):
         )
 
 
+class Roles(str, Enum):
+    ADMIN = "ADMIN"
+
+
 class ScanTargetGCP(dict):
     def __init__(self, project_id):
         dict.__init__(self, projectId=project_id)
@@ -433,90 +437,6 @@ class Client:
             )
         response.raise_for_status()
         return response
-
-    ###################################################
-    # Organization Member Invite
-    ###################################################
-
-    def iter_organization_members_invites(
-        self, organization_id: Union[UUID, str]
-    ) -> Iterator[Dict]:
-        """
-        Iterates over the members invites of an organization.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrgamizationInvites>
-        :param organization_id: the ID of the organization
-        :return: an iterator over the organization members invites objects
-        """
-        yield from self._request(
-            "GET", f"/organizations/{validate_uuid(organization_id)}/invites"
-        ).json()
-
-    def create_organization_members_invite(
-        self,
-        organization_id: Union[UUID, str],
-        email: str,
-        roles: Optional[Iterable[Roles]],
-    ) -> Iterator[Dict]:
-        """
-        Create organization member invite.
-        <https://api.zanshin.tenchisecurity.com/#operation/createOrgamizationInvite>
-        :param organization_id: the ID of the organization
-        :param email: the e-mail of the new member
-        :param roles: the Role of the member (ADMIN, None)
-        :return: a dict representing the organization member invite
-        """
-        body = {
-            "email": email,
-            "roles": roles,
-        }
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/invites",
-            body=body,
-        ).json()
-
-    def get_organization_member_invite(
-        self, organization_id: Union[UUID, str], email: str
-    ) -> Iterator[Dict]:
-        """
-        Get organization member invite.
-        <https://api.zanshin.tenchisecurity.com/#operation/getOrganizationInviteByEmail>
-        :param organization_id: the ID of the organization
-        :param email: the e-mail of the invited member
-        :return: a dict representing the organization member invite
-        """
-        return self._request(
-            "GET", f"/organizations/{validate_uuid(organization_id)}/invites/{email}"
-        ).json()
-
-    def delete_organization_member_invite(
-        self, organization_id: Union[UUID, str], email: str
-    ) -> bool:
-        """
-        Delete organization member invite.
-        <https://api.zanshin.tenchisecurity.com/#operation/deleteOrganizationInviteByEmail>
-        :param organization_id: the ID of the organization
-        :param email: the e-mail of the invited member
-        :return: a boolean if success
-        """
-        return self._request(
-            "DELETE", f"/organizations/{validate_uuid(organization_id)}/invites/{email}"
-        ).json()
-
-    def resend_organization_member_invite(
-        self, organization_id: Union[UUID, str], email: str
-    ) -> Dict:
-        """
-        Resend organization member invitation.
-        <https://api.zanshin.tenchisecurity.com/#operation/resendOrganizationInviteByEmail>
-        :param organization_id: the ID of the organization
-        :param email: the e-mail of the invited member
-        :return: a boolean if success
-        """
-        return self._request(
-            "POST",
-            f"/organizations/{validate_uuid(organization_id)}/invites/{email}/resend",
-        ).json()
 
     ###################################################
     # Organization Follower
